+++ src/apis/CommonController.php
@@ -63,6 +63,21 @@ class CommonController extends RestController
     }
     
     /**
+     * Check the status of a queue job
+     *
+     * @param integer $jobId The queue job it from the push() method
+     * @return array An array with `is_waiting`, `is_reserved` and `is_done` information.
+     */
+    public function actionQueueJob($jobId)
+    {
+        return [
+            'is_waiting' => Yii::$app->adminqueue->isWaiting($jobId),
+            'is_reserved' => Yii::$app->adminqueue->isReserved($jobId),
+            'is_done' => Yii::$app->adminqueue->isDone($jobId),
+        ];
+    }
+
+    /**
      * Get all log entries for a given scheulder model with primary key.
      *
      * @param string $model The namespace to the model